Event and buffer info cleanup
- Add virtual dtor to MemoryManagerBase
- Add event.h to arrayfire.h
- Add af_set_memory_manager_pinned, af_unset_memory_manager, and af_unset_memory_manager_pinned to Unified API
- Move buffer_info cpp api into src/api/cpp/buffer_info.cpp, and af_buffer_info_* into src/api/c/buffer_info.cpp
