+++ include/arrayfire.h
@@ -339,6 +339,7 @@ and native device id that ArrayFire is using.
 #include "af/constants.h"
 #include "af/data.h"
 #include "af/device.h"
+#include "af/event.h"
 #include "af/exception.h"
 #include "af/features.h"
 #include "af/gfor.h"
@@ -347,6 +348,7 @@ and native device id that ArrayFire is using.
 #include "af/image.h"
 #include "af/index.h"
 #include "af/lapack.h"
+#include "af/memory.h"
 #include "af/ml.h"
 #include "af/random.h"
 #include "af/seq.h"
