+++ src/api/c/CMakeLists.txt
@@ -57,6 +57,7 @@ target_sources(c_api_interface
     ${CMAKE_CURRENT_SOURCE_DIR}/bilateral.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/binary.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/blas.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/buffer_info.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/canny.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/cast.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/cholesky.cpp
