+++ src/api/c/memory.cpp
@@ -11,10 +11,10 @@
 
 #include <Array.hpp>
 #include <backend.hpp>
-#include <memory_manager.hpp>
 #include <common/err_common.hpp>
 #include <common/half.hpp>
 #include <handle.hpp>
+#include <memory_manager.hpp>
 #include <platform.hpp>
 #include <af/backend.h>
 #include <af/device.h>
@@ -27,109 +27,6 @@
 using namespace detail;
 
 using common::half;
-
-BufferInfo &getBufferInfo(const af_buffer_info handle) {
-    return *(BufferInfo *)handle;
-}
-
-af_buffer_info getHandle(BufferInfo &buf) {
-    BufferInfo *handle;
-    handle = &buf;
-    return (af_buffer_info)handle;
-}
-
-detail::Event &getEventFromBufferInfoHandle(const af_buffer_info handle) {
-    return getEvent(getBufferInfo(handle).event);
-}
-
-af_err af_create_buffer_info(af_buffer_info *handle, void *ptr,
-                             af_event event) {
-    try {
-        BufferInfo *buf = new BufferInfo({ptr, event});
-        *handle         = getHandle(*buf);
-    }
-    CATCHALL;
-
-    return AF_SUCCESS;
-}
-
-af_err af_delete_buffer_info(af_buffer_info handle) {
-    try {
-        /// NB: deleting a memory event buf does frees the associated memory
-        /// and deletes the associated event. Use unlock functions to free
-        /// resources individually
-        BufferInfo &buf = getBufferInfo(handle);
-        af_release_event(buf.event);
-        if (buf.ptr) { af_free_device(buf.ptr); }
-
-        delete (BufferInfo *)handle;
-    }
-    CATCHALL;
-
-    return AF_SUCCESS;
-}
-
-af_err af_buffer_info_get_ptr(void **ptr, af_buffer_info handle) {
-    try {
-        BufferInfo &buf = getBufferInfo(handle);
-        *ptr            = buf.ptr;
-    }
-    CATCHALL;
-
-    return AF_SUCCESS;
-}
-
-af_err af_buffer_info_get_event(af_event *event, af_buffer_info handle) {
-    try {
-        BufferInfo &buf = getBufferInfo(handle);
-        *event          = buf.event;
-    }
-    CATCHALL;
-
-    return AF_SUCCESS;
-}
-
-af_err af_buffer_info_set_ptr(af_buffer_info handle, void *ptr) {
-    try {
-        BufferInfo &buf = getBufferInfo(handle);
-        buf.ptr         = ptr;
-    }
-    CATCHALL;
-
-    return AF_SUCCESS;
-}
-
-af_err af_buffer_info_set_event(af_buffer_info handle, af_event event) {
-    try {
-        BufferInfo &buf = getBufferInfo(handle);
-        buf.event       = event;
-    }
-    CATCHALL;
-
-    return AF_SUCCESS;
-}
-
-af_err af_unlock_buffer_info_event(af_event *event, af_buffer_info handle) {
-    try {
-        af_buffer_info_get_event(event, handle);
-        BufferInfo &buf = getBufferInfo(handle);
-        buf.event       = 0;
-    }
-    CATCHALL;
-
-    return AF_SUCCESS;
-}
-
-af_err af_unlock_buffer_info_ptr(void **ptr, af_buffer_info handle) {
-    try {
-        af_buffer_info_get_ptr(ptr, handle);
-        BufferInfo &buf = getBufferInfo(handle);
-        buf.ptr         = 0;
-    }
-    CATCHALL;
-
-    return AF_SUCCESS;
-}
 
 af_err af_device_array(af_array *arr, void *data, const unsigned ndims,
                        const dim_t *const dims, const af_dtype type) {
@@ -439,8 +336,8 @@ af_err af_create_memory_manager(af_memory_manager *manager) {
     try {
         AF_CHECK(af_init());
         std::unique_ptr<MemoryManager> m(new MemoryManager());
-        MemoryManager &ref = *m.release();
-        *manager           = getHandle(ref);
+        *manager = getHandle(*m);
+        m.release();
     }
     CATCHALL;
