+++ src/api/unified/memory.cpp
@@ -48,6 +48,14 @@ af_err af_release_memory_manager(af_memory_manager handle) {
 
 af_err af_set_memory_manager(af_memory_manager handle) { return CALL(handle); }
 
+af_err af_set_memory_manager_pinned(af_memory_manager handle) {
+    return CALL(handle);
+}
+
+af_err af_unset_memory_manager() { return CALL(); }
+
+af_err af_unset_memory_manager_pinned() { return CALL(); }
+
 af_err af_memory_manager_get_payload(af_memory_manager handle, void** payload) {
     return CALL(handle, payload);
 }
